Pre-release fixes: Space key, serde cast, unwrap cleanup, docs, MSRV CI
- Fix Space key to work in both live mode and replay mode
- Fix duration serialization: explicit u64 cast to match deserialize type
- Replace guarded unwrap patterns with is_none_or()
- Update help overlay to show Space as pause/resume alias
- Update ROADMAP header to v0.14.2
- Mark FreeBSD as experimental in comparison docs
- Add MSRV CI job (Rust 1.85)

+++ src/state/session.rs
@@ -1382,7 +1382,7 @@ mod duration_serde {
     where
         S: Serializer,
     {
-        duration.as_micros().serialize(serializer)
+        (duration.as_micros() as u64).serialize(serializer)
     }
 
     pub fn deserialize<'de, D>(deserializer: D) -> Result<Duration, D::Error>
